payment pages updation

+++ orders/views.py
@@ -7,6 +7,7 @@ from decimal import Decimal
 from cart.models import CartItem
 from .models import Order, OrderItem,Perfume
 from .serializers import OrderSerializer
+from rest_framework import status
 
 
 class CreateOrderView(APIView):
@@ -20,11 +21,17 @@ class CreateOrderView(APIView):
 
         if not cart_items.exists():
             return Response({"error": "Cart is empty"}, status=400)
+        
+        payment_method = request.data.get(
+            "payment_method",
+            "COD"
+        )
+
 
         order = Order.objects.create(
             user=user,
             address=request.data.get("address"),
-            payment_method=request.data.get("payment_method", "COD")
+            payment_method=payment_method
         )
 
         total = Decimal("0.00")
@@ -36,15 +43,14 @@ class CreateOrderView(APIView):
             if perfume.stock < qty:
                 return Response(
                     {"error": f"{perfume.name} out of stock"},
-                    status=400
+                    status=status.HTTP_400_BAD_REQUEST
                 )
 
             price = perfume.price
             total += price * qty
 
-            # reduce stock
-            perfume.stock -= qty
-            perfume.save()
+            # perfume.stock -= qty
+            # perfume.save()
 
             OrderItem.objects.create(
                 order=order,
@@ -53,18 +59,35 @@ class CreateOrderView(APIView):
                 price=price
             )
 
-        order.total_amount = total
-        order.status = "PLACED"
+        order.total_amount =  total
+        
+        if payment_method =="COD":
+
+            for item in cart_items:
+                perfume=item.perfume
+                perfume.stock -= item.quantity
+                perfume.save()
+            
+            cart_items.delete()
+            order.status="PLACED"
+            order.is_paid= False
+        else:
+            order.status="PENDING"
         order.save()
 
-        # clear cart
-        cart_items.delete()
+        # order.status = "PENDING"
+        # order.save()
 
         return Response({
             "message": "Order placed successfully",
             "order_id": order.id,
             "total": total
         })
+
+        
+        # cart_items.delete()
+
+        
 
 
 class OrderHistoryView(APIView):
@@ -121,7 +144,7 @@ class BuyNowOrderView(APIView):
             price=perfume.price
         )
 
-        # reduce stock
+        
         perfume.stock -= quantity
         perfume.save()
 
@@ -134,53 +157,3 @@ class BuyNowOrderView(APIView):
             "order_id": order.id
         })
         
-
-# from django.shortcuts import render
-# from rest_framework.views import APIView
-# from rest_framework.response import Response
-# from rest_framework.permissions import IsAuthenticated
-# from .models import Order
-# from .serializers import OrderSerializer
-# from rest_framework import status
-
-
-# class CreateOrderView(APIView):
-#     permission_classes = [IsAuthenticated]
-
-#     def post(self, request):
-#         serializer = OrderSerializer(data=request.data, context={'request': request})
-
-#         if serializer.is_valid():
-#             order = serializer.save()
-#             return Response({
-#                 "message": "Order placed successfully",
-#                 "order_id": order.id
-#             })
-#         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-# class OrderHistoryView(APIView):
-#     permission_classes = [IsAuthenticated]
-
-#     def get(self, request):
-#         orders = Order.objects.filter(user=request.user).order_by('-created_at')
-#         serializer = OrderSerializer(orders, many=True)
-#         return Response(serializer.data)
-
-
-# class CancelOrderView(APIView):
-#     permission_classes = [IsAuthenticated]
-
-#     def post(self, request, order_id):
-#         try:
-#             order = Order.objects.get(id=order_id, user=request.user)
-
-#             if order.status in ["SHIPPED", "DELIVERED"]:
-#                 return Response({"error": "Cannot cancel now"}, status=400)
-
-#             order.status = "CANCELLED"
-#             order.save()
-
-#             return Response({"msg": "Order cancelled"})
-#         except Order.DoesNotExist:
-#             return Response({"error": "Not found"}, status=404)
